+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Changed
 
+- The `Tlv` and `Lv` API return `&[u8]` instead of `Option<&[u8]>`.
+
+## Added
+
+- Added `raw_data` API for `Tlv` and `Lv` to retrieve the whole `Lv`/`Tlv` slice if the object
+  was created from a raw bytestream.
+- Added `MsgToUserTlv` helper class which wraps a regular `Tlv` and adds some useful functionality.
 - `UnsignedByteField` and `GenericUnsignedByteField` `new` methods are `const` now.
 
 # [v0.7.0-beta.0] 2023-08-16